feat: alt:V update

+++ Roleplay/Factories/MyPlayerFactory.cs
@@ -1,12 +1,11 @@
 ﻿using AltV.Net;
 using AltV.Net.Elements.Entities;
-using System;
 
 namespace Roleplay.Factories
 {
     public class MyPlayerFactory : IEntityFactory<IPlayer>
     {
-        public IPlayer Create(ICore server, IntPtr playerPointer, ushort id)
+        public IPlayer Create(ICore server, nint playerPointer, uint id)
         {
             return new MyPlayer(server, playerPointer, id);
         }
