+++ Roleplay/Factories/MyVehicleFactory.cs
@@ -1,12 +1,11 @@
 ﻿using AltV.Net;
 using AltV.Net.Elements.Entities;
-using System;
 
 namespace Roleplay.Factories
 {
     public class MyVehicleFactory : IEntityFactory<IVehicle>
     {
-        public IVehicle Create(ICore server, IntPtr playerPointer, ushort id)
+        public IVehicle Create(ICore server, nint playerPointer, uint id)
         {
             return new MyVehicle(server, playerPointer, id);
         }
